Use WebApplicationInitializers from root web context to init web app
