+++ Core/src/com/infiniteautomation/mango/webapp/WebAppPackage.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright (C) 2019 Infinite Automation Software. All rights reserved.
+ */
+package com.infiniteautomation.mango.webapp;
+
+/**
+ * Package marker for the web app package
+ * @author Jared Wiltshire
+ */
+public interface WebAppPackage {
+}
