+++ Core/src/com/infiniteautomation/mango/webapp/filters/MangoForwardedHeaderFilter.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 2019 Infinite Automation Software. All rights reserved.
  */
-package com.infiniteautomation.mango.spring.webapp.filters;
+package com.infiniteautomation.mango.webapp.filters;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -13,7 +13,6 @@ import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.web.util.matcher.IpAddressMatcher;
 import org.springframework.stereotype.Component;
@@ -22,7 +21,6 @@ import org.springframework.web.filter.ForwardedHeaderFilter;
 /**
  * @author Jared Wiltshire
  */
-@Lazy
 @Component
 @WebFilter(
         filterName = MangoForwardedHeaderFilter.NAME,
